+++ geofileops/fileops.py
@@ -206,7 +206,6 @@ class LayerInfo:
         total_bounds: tuple[float, float, float, float],
         geometrycolumn: str,
         geometrytypename: str,
-        geometrytype: GeometryType,
         columns: dict[str, ColumnInfo],
         fid_column: str,
         crs: Optional[pyproj.CRS],
@@ -220,7 +219,6 @@ class LayerInfo:
             total_bounds (Tuple[float, float, float, float]): the bounds of the layer.
             geometrycolumn (str): the name of the geometry column.
             geometrytypename (str): the name of the geometry column type.
-            geometrytype (GeometryType): the type of the geometry column.
             columns (Dict[str, ColumnInfo]): the attribute columns of the layer.
             fid_column (str): the name of the fid column.
             crs (Optional[pyproj.CRS]): the crs of the layer.
@@ -231,11 +229,18 @@ class LayerInfo:
         self.total_bounds = total_bounds
         self.geometrycolumn = geometrycolumn
         self.geometrytypename = geometrytypename
-        self.geometrytype = geometrytype
         self.columns = columns
         self.fid_column = fid_column
         self.crs = crs
         self.errors = errors
+
+    @property
+    def geometrytype(self):
+        """The geometry type of the geometrycolumn."""
+        if self.geometrytypename == "NONE":
+            return None
+
+        return GeometryType(self.geometrytypename)
 
     def __repr__(self):
         """Overrides the representation property of LayerInfo."""
@@ -340,25 +345,21 @@ def get_layerinfo(
                         "An attribute column 'geometry' is not supported in a shapefile"
                     )
 
-        # Get geometry column info...
-        geometrytype = _ogr_util.ogrtype_to_geometrytype[datasource_layer.GetGeomType()]
-        if geometrytype is None:
-            geometrytypename = "NONE"
-        else:
-            # For shape files, the difference between the 'MULTI' variant and the
-            # single one doesn't exists... so always report MULTI variant by convention.
-            if driver == "ESRI Shapefile":
-                geometrytype = geometrytype.to_multitype
+        # Get geometry column name.
+        geometrytypename = _ogr_util.ogrtype_to_name(datasource_layer.GetGeomType())
 
-            assert geometrytype is not None
-            geometrytypename = geometrytype.name
+        # For shape files, the difference between the 'MULTI' variant and the
+        # single one doesn't exists... so always report MULTI variant by convention.
+        if geometrytypename != "NONE" and driver == "ESRI Shapefile":
+            if not geometrytypename.startswith("MULTI"):
+                geometrytypename = f"MULTI{geometrytypename}"
 
         # If the geometry type is not None, fill out the extra properties
         geometrycolumn = None
         extent = None
         crs = None
         total_bounds = None
-        if geometrytype is not None:
+        if geometrytypename != "NONE":
             # Geometry column name
             geometrycolumn = datasource_layer.GetGeometryColumn()
             if geometrycolumn == "":
@@ -386,7 +387,6 @@ def get_layerinfo(
                 total_bounds=total_bounds,  # type: ignore[arg-type]
                 geometrycolumn=geometrycolumn,  # type: ignore[arg-type]
                 geometrytypename=geometrytypename,
-                geometrytype=geometrytype,  # type: ignore[arg-type]
                 columns=columns,
                 fid_column=datasource_layer.GetFIDColumn(),
                 crs=crs,
@@ -1007,6 +1007,9 @@ def read_file(
     gives issues, so please report issues if they are encountered. In the future support
     for the "fiona" engine most likely will be removed. Default engine is "pyogrio".
 
+    When a file with CURVE geometries is read, they are transformed on the fly to LINEAR
+    geometries, as shapely/geopandas doesn't support CURVE geometries.
+
     Args:
         path (file path): path to the file to read from
         layer (str, optional): The layer to read. If None and there is only one layer in
@@ -1123,6 +1126,9 @@ def _read_file_base(
     """Reads a file to a pandas Dataframe."""
     # Check if the fid column needs to be read as column via the columns parameter
     fid_as_column = False
+    if isinstance(columns, str):
+        # If a string is passed, convert to list
+        columns = [columns]
     if columns is not None:
         if "fid" in [column.lower() for column in columns]:
             fid_as_column = True
@@ -2284,6 +2290,10 @@ def _append_to_nolock(
     options: dict = {},
 ):
     # Check/clean input params
+    if isinstance(columns, str):
+        # If a string is passed, convert to list
+        columns = [columns]
+
     options = _ogr_util._prepare_gdal_options(options)
     if (
         create_spatial_index is not None
@@ -2312,8 +2322,7 @@ def _append_to_nolock(
             src_layerinfo = get_layerinfo(src, src_layer, raise_on_nogeom=False)
         if (
             force_output_geometrytype is None
-            and src_layerinfo.geometrytype
-            in [GeometryType.GEOMETRY, GeometryType.GEOMETRYCOLLECTION]
+            and src_layerinfo.geometrytypename in ["GEOMETRY", "GEOMETRYCOLLECTION"]
             and not dst.exists()
         ):
             raise ValueError(
@@ -2344,13 +2353,6 @@ def _append_to_nolock(
         """
         sql_dialect = "SQLITE"
         columns = None
-
-    # When dst file doesn't exist and src is empty force_output_geometrytype should be
-    # specified, otherwise invalid output.
-    if force_output_geometrytype is None and not dst.exists():
-        if src_layerinfo is None:
-            src_layerinfo = get_layerinfo(src, src_layer, raise_on_nogeom=False)
-        force_output_geometrytype = src_layerinfo.geometrytype
 
     # Go!
     translate_info = _ogr_util.VectorTranslateInfo(
@@ -2490,7 +2492,10 @@ def copy_layer(
         explodecollections (bool, optional): True to output only simple
             geometries. Defaults to False.
         force_output_geometrytype (Union[GeometryType, str], optional): Geometry type.
-            to (try to) force the output to. Defaults to None.
+            to (try to) force the output to. In addition to geometry types, it is also
+            possible to specify PROMOTE_TO_MULTI to convert all geometries to
+            multigeometries or CONVERT_TO_LINEAR to convert CURVE geometries to linear.
+            Defaults to None.
         create_spatial_index (bool, optional): True to create a spatial index
             on the destination file/layer. If None, the default behaviour by gdal for
             that file type is respected. If the LAYER_CREATION.SPATIAL_INDEX