+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 ### Improvements
 
 - Add support for renaming column with only difference in casing (#549)
+- Add support to pass a single string for all `column` type parameters if a single
+  column should be retained (#523)
+- Enable "CURVE" geometrytype files to be processed in the general file and
+  layer operations (#558)
+- Add configuration option to only warn on dissolve errors (#561)
 
 ## 0.9.1 (2024-07-18)